backend/requirements.txt
fastapi==0.89.1
numpy==1.24.1
Pillow==9.4.0
tensorflow-gpu==2.10.1
uvicorn==0.20.0
websockets==10.4
